+++ src/components/Authentication/Kratos/ory/ui/NodeImage.tsx
